feat: make the clipboard dependency optional (#1781)
