+++ docusaurus/docs/reactnative/basics/getting_started.mdx
@@ -262,6 +262,33 @@ npm install react-native-haptic-feedback
 </TabItem>
 </Tabs>
 
+#### Copying messages
+
+Install `@react-native-clipboard/clipboard` to be able to copy messages to the clipboard.
+
+<Tabs
+  defaultValue='yarn'
+  values={[
+    { label: 'Yarn', value: 'yarn' },
+    { label: 'npm', value: 'npm' },
+  ]}
+>
+<TabItem value='yarn'>
+
+```bash
+yarn add @react-native-clipboard/clipboard
+```
+
+</TabItem>
+<TabItem value='npm'>
+
+```bash
+npm install @react-native-clipboard/clipboard
+```
+
+</TabItem>
+</Tabs>
+
 ### AndroidX Support
 
 > AndroidX is a major step forward in the Android ecosystem, and the old support library artifacts are being deprecated.
@@ -343,4 +370,12 @@ Installing this package allows you to play the video files/attachments in the ch
 
 ```bash
 expo install expo-av
+```
+
+#### Copying messages
+
+Install `expo-clipboard` to be able to copy messages to the clipboard.
+
+```bash
+expo install expo-clipboard
 ```
