+++ examples/ExpoMessaging/ios/ExpoMessaging/Supporting/Expo.plist
@@ -11,6 +11,6 @@
     <key>EXUpdatesSDKVersion</key>
     <string>46.0.0</string>
     <key>EXUpdatesURL</key>
-    <string>https://exp.host/@anonymous/ExpoMessaging</string>
+    <string>https://exp.host/@madsroskar/ExpoMessaging</string>
   </dict>
 </plist>
