+++ examples/ExpoMessaging/package.json
@@ -11,9 +11,10 @@
     "@react-native-masked-view/masked-view": "0.2.7",
     "@react-navigation/native": "5.8.10",
     "@react-navigation/stack": "5.12.8",
-    "expo": "~46.0.9",
+    "expo": "~46.0.16",
     "expo-av": "~12.0.4",
     "expo-blur": "~11.2.0",
+    "expo-clipboard": "^3.1.0",
     "expo-document-picker": "~10.3.0",
     "expo-file-system": "~14.1.0",
     "expo-haptics": "~11.3.0",
@@ -25,7 +26,7 @@
     "expo-status-bar": "~1.4.0",
     "react": "18.0.0",
     "react-dom": "18.0.0",
-    "react-native": "0.69.4",
+    "react-native": "0.69.6",
     "react-native-gesture-handler": "~2.5.0",
     "react-native-reanimated": "~2.9.1",
     "react-native-safe-area-context": "4.3.1",
