+++ package/expo-package/src/handlers/index.ts
@@ -1,13 +1,13 @@
-export * from './shareImage';
-export * from './deleteFile';
 export * from './compressImage';
+export * from './deleteFile';
 export * from './getLocalAssetUri';
 export * from './getPhotos';
 export * from './NetInfo';
 export * from './pickDocument';
 export * from './saveFile';
 export * from './setClipboardString';
+export * from './shareImage';
+export * from './Sound';
 export * from './takePhoto';
 export * from './triggerHaptic';
-export * from './Sound';
 export * from './Video';
