+++ package/native-package/package.json
@@ -31,15 +31,18 @@
   },
   "devDependencies": {
     "@react-native-camera-roll/camera-roll": "^5.0.2",
+    "@react-native-community/netinfo": ">=2.0.7",
+    "@stream-io/flat-list-mvcp": "0.10.2",
     "@types/react-native-share": "^3.3.3",
+    "react-native": ">=0.60.0",
     "react-native-document-picker": "^8.1.1",
+    "react-native-fs": ">=2.16.6",
     "react-native-haptic-feedback": "^1.14.0",
     "react-native-image-crop-picker": "^0.38.0",
-    "@stream-io/flat-list-mvcp": "0.10.2",
-    "@react-native-community/netinfo": ">=2.0.7",
-    "react-native": ">=0.60.0",
-    "react-native-fs": ">=2.16.6",
     "react-native-image-resizer": ">=1.4.2",
     "react-native-share": ">=4.1.0"
+  },
+  "optionalDependencies": {
+    "@react-native-clipboard/clipboard": "^1.11.1"
   }
 }
