+++ package/native-package/src/index.js
@@ -11,6 +11,7 @@ import {
   getPhotos,
   pickDocument,
   saveFile,
+  setClipboardString,
   shareImage,
   Sound,
   takePhoto,
@@ -27,6 +28,7 @@ registerNativeHandlers({
   NetInfo,
   pickDocument,
   saveFile,
+  setClipboardString,
   SDK: 'stream-chat-react-native',
   shareImage,
   Sound,
